Feature model pipeline (#10)
* Implementing pipeline models and noiseinjector component

* Cleaning and restructuring folders

* Ensures model is marked as fitted

Sets the `fitted_` attribute to True after the model is trained. This ensures that subsequent calls to `predict_proba` will not attempt to use an untrained model.

Casts the `y_batch` to int8 to avoid potential overflow issues.

* Refactors validation data handling in models

Moves validation data to the base models for RNN, LSTM, and GRU, ensuring consistent handling and usage during training and evaluation. This change also updates the model instantiation to correctly pass the validation data.

* Uses score_samples for prediction

Updates the `predict` method in the base models to use `score_samples` for making predictions.